+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `:vips` will be documented in this file.
 ### Added
 - fix minor formatting issues reported by phpcs [John Cupitt]
 - regenerated autodocs from libvips 8.4 to reduce confusion [John Cupitt]
+- add Image::hasAlpha tester
 
 ### Deprecated
 - Nothing